Modify the Graph component the minimum possible to get it working as a lumino widget and standalone vue component
