+++ src/styles/cylc/_workflow.scss
@@ -37,7 +37,7 @@
         }
 
         .p-TabBar-tabCloseIcon:before {
-          content: '🗙';
+          content: 'X';
         }
       }
     }
